Fix usage of stdarg macros.
A variable argument retrieved with va_start() can only be used
in a function argument once, on some platforms. After the call,
its value is indeterminate and it must be destroyed using
va_end(), and a new copy be retrieved by redoing va_start().

This fixes a segmentation fault when writing to the log file.